+++ src/app/core/services/generator.factory.ts
@@ -1,0 +1,9 @@
+import { InjectionToken } from "@angular/core";
+
+import { GeneratorService } from "./generator";
+
+export const GeneratorToken5 = new InjectionToken<string>("GeneratorToken5");
+
+export function GeneratorFactory(n: number){
+    return (generator: GeneratorService) => generator.generate(n);
+}